+++ test/functional/rpc_signrawtransaction.py
@@ -12,6 +12,7 @@ class SignRawTransactionsTest(BitcoinTestFramework):
     def set_test_params(self):
         self.setup_clean_chain = True
         self.num_nodes = 1
+        self.extra_args = [["-deprecatedrpc=signrawtransaction"]]
 
     def successful_signing_test(self):
         """Creates and signs a valid raw transaction with one input.